PSM Interop: add support to enable CSM Observability and a new test case (#18)
This PR is identical to https://github.com/grpc/grpc/pull/34835, but was
created before this repo was forked.

---

This PR adds CSM Observability testing capability in the PSM Interop
testing framework. This PR mostly changes the framework Python code.

This adds a flag `enable_csm_observability` to the client / server
deployment yaml file such that, when enabled, we will create a GMP
`PodMonitoring` resource and pass the `--enable_csm_observability` to
each language's client / server container (for them to actually
enable the Prometheus endpoint)

I added a new test under `tests/csm/csm_observability_test.py`. This
is basically a copy of the `tests/baseline_test.py` but with the
`enable_csm_observability=True`. 

Other PRs for this whole thing to work:
- https://github.com/grpc/grpc/pull/34752: The `PodMonitoring`
  resource yaml template
- https://github.com/grpc/grpc/pull/34832: Support for the
  `--enable_csm_observability` flag in the C++ client/server image

---------
Co-authored-by: Sergii Tkachenko <sergiitk@google.com>