Paramatize request/response payload size (#20)
Currently, the xds interop client is sending an empty payload for the
`UnaryCall` rpc. Since testing CSM Observability will need a non-trivial
payload (e.g. to verify the metrics about the number of bytes being
sent), this PR will try to add 2 parameters `request_payload_size` and
`response_payload_size` to the client image to control the size of the
payload to be sent. Each language will need to implement this parameter
in their respective client image. The one for c++ is
[here](https://github.com/grpc/grpc/pull/35545).

Note that we only need to set these 2 parameters (both `request` and
`resposne` payload size) on the _client deployment_ because the client
will in turn tell the server how much to send back.

The good 'ol interop client already has these:
https://github.com/grpc/grpc/blob/master/test/cpp/interop/interop_client.cc#L249-L251