Style: add typing-extensions per updated Google Python Style Guide  (#45)
[Google Python Style
Guide](https://google.github.io/styleguide/pyguide.html) has evolved,
and looks all modern and futuristic now. I'll be getting tolling in this
repo closer to match closer to it.

- Add
[`typing-extensions`](https://typing-extensions.readthedocs.io/en/latest/)
package.
- Configure isort single_line_exclusions according to
https://google.github.io/styleguide/pyguide.html#2241-exemptions

With `typing-extensions`, we can use such cool things as
[`TypeAlias`](https://docs.python.org/3/library/typing.html#typing.TypeAlias),
[`Self`](https://docs.python.org/3/library/typing.html#typing.Self), and
[`@override`](https://docs.python.org/3/library/typing.html#typing.override).

Note: we don't migrate the code to the new types yet, but we should be
using these features when writing new code.