Fixes new pod log collection when runner.run() called multiple times (#58)
In the subsetting test, only the first client's log was properly
collected. The rest were virtually empty files, with only `Finished log
collection for pod psm-grpc-client-..` in them.

This fixes an issue logs not being collected for subsequent calls to
client/server runner's `#run()`.
The bug was in the shared log stop threading.Event being reused between
different runs. So after the first run sets, log collectors for the
following runs see that it's set and immediately finish collecting.