csm: Pass value to argument --enable_csm_observability
This is needed by the Java client/server, because there's a simplistic
parsing implementation. Since we need the parsing to work
cross-language, we always include a value. `--secure_mode` and
`--fail_on_failed_rpcs` (unused by this framework, but defined in
xds-test-descriptions.md) are both boolean flags that we pass a value.