framework core: fixes a case when there are fewer endpoints than backends (#93)
Fixes an issue with NEG size and Backend Health waiting logic.

This bug has been in the framework from the very beginning, because we
never tested the flow with (a) Regional or multizonal cluster, AND (b) a
single server.

The issue: when GKE cluster has nodes in more than one zone, a NEG for
the test server/service is created in each zone. If there's fewer
endpoints than zones in the cluster, not every NEG will end up having a
server endpoint. Therefore:

1. `wait_for_network_endpoint_group` will fail because it waits for
every NEG to have more than 0 endpoints.
2. `wait_for_backends_healthy_status` will fail because NEGs with 0
endpoints will never be marked as healthy.

This PR changes the logic of:
1. `wait_for_network_endpoint_group` now waits for a NEG to merely exist
in GCP APIs.
2. `wait_for_backends_healthy_status` now accounts for server
replica_count and will succeeded as soon as the number of healthy
Backends to be greater or equal the number of server replicas.

Discovered when debugging #86. Splitting to its own PR.