Increased retry time for gamma.affinity_session_drain_test (#158)
failing stage: 04_confirm_all_servers_receive_traffic

We are verifying all 3 server receives traffic. Channel becomes ready as
soon as one of the server is connected. In this case we see all RPCs
going to that. We need to give time for other instances to come up.

Increasing the retry timeout from 1 min to 2 min to reduce flakiness.

Internal: b/379040924