Dualstack test: fix the order of server namespace cleanup. (#170)
Previously, `force_namespace=True` was passed to
`self.v6_server_runner.cleanup()`, but the the call to this method is
conditional on `if self.v6_server_runner:`.

Therefore, when `self.v6_server_runner` isn't set, there won't be
anything to call forced namespace cleanup. In practice, I don't see this
happening as the code is right now, but still fixing the code for
future-proofing (f.e. bad refactoring or copy-pasting to a different
code flow).