[affinty_test] Add retry logic for asserting eds length (#171)
The affinity_test is failing because server backends take time to ramp
up to 3 or the number of replicas required , so when we request the
status, we get only 1 endpoint. Adding a retry logic will give some time
for backends to ramp up and only fail after a fixed timeout - 10 mins.

Bug : b/373265335