Reorder checks for RPC status codes (#181)
Earlier we were checking unwanted status and then the correct expected
code.
Reordering this to check if we get expected status and then check for
any unwanted status.

Reason: We are seeing a Flake where we get unwanted status b/416171553 .
We wanted to create the matcher to ignore unwanted status, hence it make
sense to swap checks as mentioned in cl/758143043