[GAMMA] Increase Affinity Session Drain test resilience: skip step and bump replicas (#178)
Removed the step ""04_confirm_all_servers_receive_traffic" . 

Issue: sometimes all servers don't receive traffic as some are created
in another zone with priority 1

This test will be able to check session affinity without this step as we
assert on number of RPC after setting Cookie.
Also increasing replica count so we have more chances of getting greater
than 1 server in a subzone.

Only following case will be able to skip test.
It is possible that only 1 server is created in priority 0 . In this
case our test would still pass but it will not confirm the feature is
working correctly.

Internal b/379040924