[Circuit Breaking test] Increase RPCs in flight error threshold to 5 from 1 (#186)
The number of failures exceeding this 1% threshold has increased
recently, where most failures are just slightly off the threshold. Hence
increasing the threshold should help reduce these failures.

Some examples of recently failing errors which would pass with a 2%
threshold:
```
"790" unexpectedly not between "792" and "808" : Found wrong number of RPCs in flight: actual(790), expected(800 ± 1%)
"984" unexpectedly not between "990" and "1010" : Found wrong number of RPCs in flight: actual(984), expected(1000 ± 1%)
"791" unexpectedly not between "792" and "808" : Found wrong number of RPCs in flight: actual(791), expected(800 ± 1%)
```

There are also a few which would pass only with a 5% threshold.
```
"524" unexpectedly not between "495" and "505" : Found wrong number of RPCs in flight: actual(524), expected(500 ± 1%)
"766" unexpectedly not between "792" and "808" : Found wrong number of RPCs in flight: actual(766), expected(800 ± 1%)
"769" unexpectedly not between "792" and "808" : Found wrong number of RPCs in flight: actual(769), expected(800 ± 1%)
```

Taking the maximum of these, have increased the threshold to 5.