Implement a federation test (#185)
This uses some of the tools that were originally implemented for the
fallback test to implement a test for the [federation
feature](https://github.com/grpc/proposal/blob/master/A47-xds-federation.md).
The test spreads resources across two control plane servers, so that
both a client and a server need to use the authority information in the
resource URLs to know which control plane server to request each
resource from, and they need to contact both control plane servers to
get their complete configurations. go-control-plane imposes some
constraints on what combinations of resources it will serve, so both
control plane servers have all of the server resources, but the
specified authorities should cause the server to contact both anyway,
and we attempt to confirm that by testing that it has a connection to
both control plane servers.

While implementing this, I had to make a couple of changes that impacted
the fallback test:

- I modified the bootstrap file template to include the federation
fields, and I made the bootstrap generator function more general.
- I modified the resource generation code to be able to generate
individual resources and group them together separately.

Test runs:

- [ ]
[grpc/java/master/branch/psm-light](https://source.cloud.google.com/results/invocations/c31fba7f-6afc-40b1-9c54-70bbd3b854fc)
(the fallback test failed, but it is also failing the same way without
this change)
- [x]
[grpc/go/master/branch/psm-light](https://source.cloud.google.com/results/invocations/5af69c83-1eb2-4c20-9d88-88573fce2e60)
- [x]
[grpc/core/master/linux/psm-light](https://source.cloud.google.com/results/invocations/2c3d4e24-32d1-4b61-93e4-bb60712e071a)
- [x]
[grpc/core/master/linux/psm-light-python](https://source.cloud.google.com/results/invocations/b6a9933f-4268-4f43-b86a-8e81224b7a3f)
- [x]
[grpc/node/master/psm-light](https://source.cloud.google.com/results/invocations/4d02b36e-3928-4975-936e-22232709d205)